Response to reviewer
@jj Please check:
6) You still did not write at all about the results, we see in the
plots and tables. Here is our old remark: There is no discussion about
the results. A reader has to look into all data to get an idea of how
the algorithm performs. Here, the authors should give some sort of
summary of what can be seen in the data and present the most important
findings of the authors. Where is the algorithm competitive? Where does
the algorithm fail?